requirements.txt
openai>=1.0.0
python-dotenv
pyautogen
autogen-agentchat
autogen-ext
autogen-ext[mcp]
pdfplumber
python-docx
requests
fastapi
websockets
streamlit
psycopg2-binary
azure-core
azure-search-documents
autogen-core 
aiohttp
azure-identity